[PORT FROM MAIN] WiFi: "Connect" option disappear after rotating.
BZ: 34109

The connect option disappears because there is a problem in saving and
restoring the state of the access point showed in the dialog box. When a
rotation accurs, the framework calls onSaveInstanceState(Bundle outState)
of the wifisettings activity. In this method, the state of the access is
saved, but not its rssi level. After a rotation, when restoring the access
point state, the rssi is initialized to the default value, which means out of
range. This is why the connect option is not displayed. I fixed the issue by
saving and restoring the rssi level of the access point.

Feature : WiFi
Category : AOSP improvement
Severity: medium
Origin : internal
Upstream-candidate: yes

Change-Id: Ibd58132c922828941603673383ff81f25d61556e
Orig-Change-Id: I624cc8ce82a69fd94931088dc0f7833663ef0772
Signed-off-by: Paul Naoumenko <paulx.naoumenko@intel.com>
Signed-off-by: Jean-Marie Lemetayer <jean-mariex.lemetayer@intel.com>
Reviewed-on: http://android.intel.com:8080/121980
Reviewed-by: Poulain, LoicX <loicx.poulain@intel.com>
Reviewed-by: Bachot, Jean-michel <jean-michel.bachot@intel.com>
Tested-by: Champciaux, Nicolas <nicolas.champciaux@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>