Wifi Settings: Saved AP were displayed as "not in range" before first scan.
BZ: 69798

In Wifi Settings, immediately after turning ON the Wifi, saved access points
are displayed as "not in range" while no scan has been completed yet. This is
not coherent since when starting Wifi, before the first scan, we d'ont know
anything about these AP so we don't have to display any status about then.

To solve this, we don't display saved access points prior to first scan result.
After the first the scan, either saved access points are in the range and they
are displayed normally, or they are not in th range and they are added at the
end of the list as "not in range"

Change-Id: I3d48d1b808ce47c4828cfba8a870992de9fecf81
Signed-off-by: Matthieu Mauger <matthieux.mauger@intel.com>
Reviewed-on: http://android.intel.com:8080/80616
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Reviewed-by: Bachot, Jean-michel <jean-michel.bachot@intel.com>
Tested-by: Zurmely, PierreX <pierrex.zurmely@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>