[PORT FROM MAIN]wifi: Saved AP were displayed as "not in range" before first scan.
BZ: 69798

In Wifi Settings, immediately after turning ON the Wifi, saved access points
are displayed as "not in range" while no scan has been completed yet. This is
not coherent since when starting Wifi, before the first scan, we d'ont know
anything about these AP so we don't have to display any status about then.

To solve this, we don't display saved access points prior to first scan result.
After the first the scan, either saved access points are in the range and they
are displayed normally, or they are not in th range and they are added at the
end of the list as "not in range"

Change-Id: I9844815de8ed4694437273ded35c313ba9f5ac81
Orig-Change-Id: I1d6e35b82ca9a3365c7b0b7b77e8efe0fa17f6b9
Category: device enablement
Domain: CWS.WIFI-Common
Origin: internal
Upstream-Candidate: yes
Signed-off-by: Matthieu Mauger <matthieux.mauger@intel.com>
Reviewed-on: http://android.intel.com:8080/121988
Reviewed-by: Poulain, LoicX <loicx.poulain@intel.com>
Reviewed-by: Bachot, Jean-michel <jean-michel.bachot@intel.com>
Reviewed-by: Naoumenko, PaulX <paulx.naoumenko@intel.com>
Tested-by: Champciaux, Nicolas <nicolas.champciaux@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>