WLAN: Set SIM only as eap method when SIM is selected
BZ: 70399

Currently, when SIM is selected as EAP method,  both SIM and AKA
are set as eap method in network configuration (wpa_supplicant.conf).

In this case, the supplicant try first AKA method (in fact provide
AKA identity : 0+IMSI) because it's the more secure method.
If the radius server is not well configured this could lead to
a connection failure. This is the case of freeradius server.

As the GUI selector provide 2 selector, one for SIM and an other for AKA,
, there is no reason to set SIM AKA when SIM is selected on GUI.

In this case the supplicant will directly reply with SIM identity
(1+IMSI) when EAP-Request-Identity frame is received.

Change-Id: Id6d2c82b60a23b1c8902e1c37305e409ac448233
Orig-Change-Id: I8180c1368367413a9e5a3c3412bd759090793c55
Signed-off-by: Honore Tricot <honorex.tricot@intel.com>
Reviewed-on: http://android.intel.com:8080/83504
Reviewed-by: Naoumenko, PaulX <paulx.naoumenko@intel.com>
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Tested-by: Champciaux, Nicolas <nicolas.champciaux@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>