[PORT FROM MAIN]Wifi Settings: "not in range" was displayed during WPS process
BZ: 71438

When launching WPS, if there is no AP in the range or only a few
ones but not enough to fill the screen, an additional ghost AP
can be seen at the end of the list. This AP hasn't SSID and only
a message “not in range” is set as summary.

In fact, in the WifiConfigStore class, when the WPS starts, it
creates a temporary dummy BSS for its internal needs.
However this BSS is included into the list of configured networks
and it is returned by the getConfiguredNetworks() function.

Since it is too risky to change deeply the WPS mechanism, because
WPS also uses the getConfiguredNetworks() function, it was preferred
to work at an upper layer, in the WifiSettings class.
In this class, when constructing the AccessPoints list from the
return of getConfiguredNetworks(), BSSs with a null or empty SSID
are discarded.

Change-Id: Id7077c8bd23b7750dc4472e2719a0863ce69e96b
Orig-Change-Id: I0a700e0a7a997d50ce4c05d2c00c71ef795547e9
Category: device enablement
Domain: CWS.WIFI-Common
Origin: internal
Upstream-Candidate: yes
Signed-off-by: Matthieu Mauger <matthieux.mauger@intel.com>
Reviewed-on: http://android.intel.com:8080/121990
Reviewed-by: Poulain, LoicX <loicx.poulain@intel.com>
Reviewed-by: Bachot, Jean-michel <jean-michel.bachot@intel.com>
Reviewed-by: Naoumenko, PaulX <paulx.naoumenko@intel.com>
Tested-by: Champciaux, Nicolas <nicolas.champciaux@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>