[PORT FROM JBMR0] Settings: Disable SIM lock UIs in Airplane mode
BZ: 58530

If the airplane mode is activated using the power
button and if the current activity is one of the
SIM lock activities, then the SIM lock UI is not
disabled. Since the framework doesn't completely
support SIM access in airplane mode, there are few
issues seen.

This patch disables/greys out the SIM lock UIs when
the airplane mode is activated using the power button.

Change-Id: I45fe982d7a35e4a83ab04786101d1f8c99149a34
Orig-Change-Id: I6e9538a7c77afbf1043515fc708b7b5e41d57a1d
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>
Signed-off-by: Arun Ravindran <arun.ravindran@intel.com>