[PORT FROM MAIN] Wifi Settings: set default IP and Proxy methods for open AP
BZ: 80255

When connecting to a secured AP for the first time, the WifiDialog
is shown and the IP and Proxy methods are set according to the
dialog spinners.
When connecting to an open AP, the WifiDialog is skip and thus IP
and Proxy methods are never set.

To solve this, the IP and Proxy methods are set to default
(respectively DHCP and NONE) when selecting an open AP to connect to
for the first time. These default values are the same as done by the
WifiDialog for secured AP.

Category: AOSP improvement
Domain: WIFI-Settings
Severity: medium
Origin: internal
Upstream-Candidate: no

Change-Id: I0f2ec6d8dbe0538140c1e630c948cdc88037b66f
Orig-Change-Id: I6f884a4f0122d58dc5d5a640afd0653e1853aec5
Signed-off-by: Matthieu Mauger <matthieux.mauger@intel.com>
Reviewed-on: http://android.intel.com:8080/122755
Reviewed-by: Poulain, LoicX <loicx.poulain@intel.com>
Reviewed-by: Naoumenko, PaulX <paulx.naoumenko@intel.com>
Tested-by: Champciaux, Nicolas <nicolas.champciaux@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>