Wifi Settings: set default IP and Proxy methods for open AP
BZ: 80255

When connecting to a secured AP for the first time, the WifiDialog
is shown and the IP and Proxy methods are set according to the
dialog spinners.
When connecting to an open AP, the WifiDialog is skip and thus IP
and Proxy methods are never set.

To solve this, the IP and Proxy methods are set to default
(respectively DHCP and NONE) when selecting an open AP to connect to
for the first time. These default values are the same as done by the
WifiDialog for secured AP.

Change-Id: I6f884a4f0122d58dc5d5a640afd0653e1853aec5
Signed-off-by: Matthieu Mauger <matthieux.mauger@intel.com>
Reviewed-on: http://android.intel.com:8080/87938
Reviewed-by: Naoumenko, PaulX <paulx.naoumenko@intel.com>
Reviewed-by: Chen, HuaqiangX <huaqiangx.chen@intel.com>
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Reviewed-by: De Cesco, Jonathan <jonathan.de.cesco@intel.com>
Tested-by: Zurmely, PierreX <pierrex.zurmely@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>