[PORT FROM MAIN]wifi: set default IP and Proxy methods for open AP
BZ: 80255

When connecting to a secured AP for the first time, the WifiDialog
is shown and the IP and Proxy methods are set according to the
dialog spinners.
When connecting to an open AP, the WifiDialog is skip and thus IP
and Proxy methods are never set.

To solve this, the IP and Proxy methods are set to default
(respectively DHCP and NONE) when selecting an open AP to connect to
for the first time. These default values are the same as done by the
WifiDialog for secured AP.

Category: device enablement
Domain: CWS.WIFI-Common
Origin: internal
Upstream-Candidate: yes

Change-Id: Ia817d6f8c0e54658dff605244253709992d0d8af
Orig-Change-Id: I0f2ec6d8dbe0538140c1e630c948cdc88037b66f
Signed-off-by: Matthieu Mauger <matthieux.mauger@intel.com>
Reviewed-on: http://android.intel.com:8080/122755
Reviewed-by: Poulain, LoicX <loicx.poulain@intel.com>
Reviewed-by: Naoumenko, PaulX <paulx.naoumenko@intel.com>
Tested-by: Champciaux, Nicolas <nicolas.champciaux@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>