[PORT FROM MAIN]wifi: Smart sorting of access points in WiFi list
BZ: 88749 92584

Until now, access points in WiFi list was sorted according to their RSSI.
However the sorting process was extremely sensitive leading to a lot of APs
reorganization trigged by RSSI changes. This gives a boring user experience
because APs are permanently rearranged and it's difficult to browse the list and
select the desired one.

A first solution could be to sort APs in alphabetic order, but it most case user
prefers to have APs sorted by quality because this is the most important
criteria to select an AP to connect to.

From a technical point of view, the list was populated by AccessPoint objects
which were completely destroyed and re-created during each refresh, this was
time and resources consuming.

The aim of this patch is to maintain the list of AccessPoint objects and just
update them when a minor change occurs like a change of RSSI. The position of
each AccessPoint in the list still depends of the RSSI but the reorganization is
now triggered using an hysteresis algorithm. Thanks to the hysteresis, the list
is now stable when DUT stays at the same place but it's still updated quickly
when DUT is moving.

For important events like when a configured AP is added of forgotten, the list
is completely cleared and rebuilt as it was done before. This is to keep a high
reliability of Settings app.

From this smart sorting, it is resulting a better user experience. The UI now
offers a stable list of APs sorted by quality and it's easier for the user to
select the suitable AP.

Category: device enablement
Domain: CWS.WIFI-Common
Origin: internal
Upstream-Candidate: yes

Change-Id: I7d2c972f3e1326ec2efef69e3f3e7db3ef7ae33d
Orig-Change-Id: Ibe112f022db02b075ba1a8832743b80687189b15
Signed-off-by: Matthieu Mauger <matthieux.mauger@intel.com>
Reviewed-on: http://android.intel.com:8080/122242
Reviewed-by: Poulain, LoicX <loicx.poulain@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Reviewed-by: Naoumenko, PaulX <paulx.naoumenko@intel.com>
Tested-by: Champciaux, Nicolas <nicolas.champciaux@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>