[PORT FROM R42S] WiFi Settings: shows hotspot vibration option whatever the security
BZ: 98400

The recent check-box option “user option to enable vibration when hotspot ready”
(patch 96752) has been misplaced in the security group. This makes the option
hidden when Hotspot is configured with no security.

This patch moves the check-box from security group to general group allowing to
see it whatever the chosen security.

Change-Id: I3fb1f3fcb7639bf1c7ddaf4ced675484ab296468
Signed-off-by: Matthieu Mauger <matthieux.mauger@intel.com>
Reviewed-on: http://android.intel.com:8080/100514
Reviewed-by: Samoun, JacquesX <jacquesx.samoun@intel.com>
Reviewed-by: Bachot, Jean-michel <jean-michel.bachot@intel.com>
Reviewed-by: Chen, HuaqiangX <huaqiangx.chen@intel.com>
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Reviewed-by: Trivelly, Jean <jean.trivelly@intel.com>
Tested-by: Zurmely, PierreX <pierrex.zurmely@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>