wifi: enable button for switching wireless display on/off for Baylake
BZ: 100939

On Baylake, the UI layout used is the tablet one.

However, the native "setings" apk from AOSP seems to have been coded so that we cannot enable Wireless Display for tablets.

The commit comment from AOSP seems to go this way:

   <<<<<
    First draft of wifi display settings UI.
    All of the functionality is in place but the UI is
    not fully polished yet.
   >>>>>>

This patch enables the wireless display on/off button for tablets.
This must be considered as a workaround waiting for the UI to natively support it.
It allows to quickly unlock Baylake program.

Change-Id: Idadd2802cbdf0bf861af78ef5effe94cfb91ef20
Signed-off-by: Jeremie GARCIA <jeremiex.garcia@intel.com>
Reviewed-on: http://android.intel.com:8080/102229
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Zurmely, PierreX <pierrex.zurmely@intel.com>
Tested-by: Zurmely, PierreX <pierrex.zurmely@intel.com>
Reviewed-by: Chen, HuaqiangX <huaqiangx.chen@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>