Settings: SIM Card Option doesn't reappear when SIM is inserted
BZ: 106334

When a SIM PIN enabled SIM is removed and inserted while in setup SIM Card lock
screen,after entering the PIN the option doesn't reappear.

The reason for this is, when the SIM is removed while in Setup SIM Card Lock screen
the screen is dimmed, when the user presses back the resume of Security Settings is
called and since there is NO SIM the SIM LOCK option is removed. But when the SIM
is re-inserted, the updatesimpreferences, finds that there is no SIM PIN option.

This patch fixes this issue by calling the createPreferenceHierarchy, if there is
no SIM LOCK option and the device has an Icc card.

Change-Id: Iaccf8b73f2fa6617271c458a5897155f4d09e579
Signed-off-by: Arun Ravindran <arun.ravindran@intel.com>
Reviewed-on: http://android.intel.com:8080/109771
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Lucas, Guillaume <guillaume.lucas@intel.com>
Reviewed-by: Garipoli, Giuseppe <giuseppe.garipoli@intel.com>
Reviewed-by: Fremaux, Pascal <pascal.fremaux@intel.com>
Reviewed-by: Chen, HuaqiangX <huaqiangx.chen@intel.com>
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Reviewed-by: Badrappan, Jeevaka <jeevaka.badrappan@intel.com>
Reviewed-by: Louis, NicolasX <nicolasx.louis@intel.com>
Tested-by: Louis, NicolasX <nicolasx.louis@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>