[PORT FROM MAIN] WiFi Settings: Load configured networks asynchronously
BZ: 97532

The refresh of WiFi list needs to retrieve configured networks. This task
requires an access to the file system and sometimes this can take a moment.
In some rare case it asks so much time that it trigs an ANR of Settings
application.

To solve this, the access to file is now done in an AsyncTask to not block the
UI thread.

Category: AOSP improvement
Domain: WIFI-Settings
Severity: medium
Origin: internal
Upstream-Candidate: yes

Change-Id: I614928a68c935236349b897be460c8d2e12a368e
Orig-Change-Id: I9e0e956c0bf3dfbb563d7d6a068dc7bb3601b533
Signed-off-by: Matthieu Mauger <matthieux.mauger@intel.com>
Reviewed-on: http://android.intel.com:8080/122386
Reviewed-by: Poulain, LoicX <loicx.poulain@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Reviewed-by: Naoumenko, PaulX <paulx.naoumenko@intel.com>
Tested-by: Champciaux, Nicolas <nicolas.champciaux@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>