[PORT FROM R42S] WiFi Settings: Load configured networks asynchronously
BZ: 97532

The refresh of WiFi list needs to retrieve configured networks. This task
requires an access to the file system and sometimes this can take a moment.
In some rare case it asks so much time that it trigs an ANR of Settings
application.

To solve this, the access to file is now done in an AsyncTask to not block the
UI thread.

Change-Id: I9e0e956c0bf3dfbb563d7d6a068dc7bb3601b533
Signed-off-by: Matthieu Mauger <matthieux.mauger@intel.com>
Reviewed-on: http://android.intel.com:8080/110516
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Chen, HuaqiangX <huaqiangx.chen@intel.com>
Reviewed-by: Bachot, Jean-michel <jean-michel.bachot@intel.com>
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Tested-by: Zurmely, PierreX <pierrex.zurmely@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>