[PORT FROM MAIN] WiFi Settings: Save connection DetailedState before screen rotation
BZ: 110273

In WiFi Settings, when a screen rotation is done while a popup is displayed in
front, the connection state of current AP shown in background was not coherent.

This patch saves the connection state in order to update the display immediately
after screen rotation without waiting for event which are not handled until the
popup is closed.

Category : AOSP Improvement
Domain: WIFI-Settings
Origin: internal
Upstream-Candidate: yes

Change-Id: I53877c5f55ef7406ca556a1b7517ff80d9d3ef8f
Orig-Change-Id: I948031dd4dfd7521cd53ab014b0612ebc287b5fd
Signed-off-by: Matthieu Mauger <matthieux.mauger@intel.com>
Reviewed-on: http://android.intel.com:8080/125007
Reviewed-by: Poulain, LoicX <loicx.poulain@intel.com>
Reviewed-by: Naoumenko, PaulX <paulx.naoumenko@intel.com>
Tested-by: Champciaux, Nicolas <nicolas.champciaux@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>