WiFi Settings: Save connection DetailedState before screen rotation
BZ: 110273

In WiFi Settings, when a screen rotation is done while a popup is displayed in
front, the connection state of current AP shown in background was not coherent.

This patch saves the connection state in order to update the display immediately
after screen rotation without waiting for event which are not handled until the
popup is closed.

Change-Id: I948031dd4dfd7521cd53ab014b0612ebc287b5fd
Signed-off-by: Matthieu Mauger <matthieux.mauger@intel.com>
Reviewed-on: http://android.intel.com:8080/111528
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Naoumenko, PaulX <paulx.naoumenko@intel.com>
Reviewed-by: Becle, MichelX <michelx.becle@intel.com>
Reviewed-by: Chen, HuaqiangX <huaqiangx.chen@intel.com>
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Tested-by: Zurmely, PierreX <pierrex.zurmely@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>