WiFi Display: Adapter reconnection upon AP just connected on another frequency
BZ: 113453

Since Broadcom chip has some limitation to offer high video quality while used
in muli-role and multi-channel, the best solution is to use the same channel as
the access point when running in multi-role.

This solution is already in place when DUT is first connected to an access point
and then connected to a video adapter. However, in some case we can image that
user first connects to video adapter and then connects to an access point. In
such situation, there is a  high probability that adapter and access point
doesn't run on the same channel.

This patch detects when DUT is reaching an access point with a different
frequency than video adapter, a dialog box informs the user that the video link
will be closed and restored for better quality. The user has no other choice
than accept and the patch manages the disconnection, retrieves the adapter we
left and finally performs the reconnection.

Change-Id: I72fb7ca84382cb246779a7eb4ef0b9ebe01b95d0
Signed-off-by: Matthieu Mauger <matthieux.mauger@intel.com>
Reviewed-on: http://android.intel.com:8080/113235
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Trivelly, Jean <jean.trivelly@intel.com>
Reviewed-by: Chen, HuaqiangX <huaqiangx.chen@intel.com>
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Reviewed-by: Samoun, JacquesX <jacquesx.samoun@intel.com>
Reviewed-by: Bachot, Jean-michel <jean-michel.bachot@intel.com>
Tested-by: Lestienne, ChristopheX <christophex.lestienne@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>