[PORT FROM MAIN JB-MR2]widi: Adapter reconnection upon AP just connected on another frequency
BZ: 113453 122779 123309

Since Broadcom chip has some limitation to offer high video quality while used
in muli-role and multi-channel, the best solution is to use the same channel as
the access point when running in multi-role.

This solution is already in place when DUT is first connected to an access point
and then connected to a video adapter. However, in some case we can image that
user first connects to video adapter and then connects to an access point. In
such situation, there is a  high probability that adapter and access point
doesn't run on the same channel.

This patch detects when DUT is reaching an access point with a different
frequency than video adapter, a dialog box informs the user that the video link
will be closed and restored for better quality. The user has no other choice
than accept and the patch manages the disconnection, retrieves the adapter we
left and finally performs the reconnection.

Closes the Wireless Display Setting
window as soon as the reconnection process is completed. By the way, the user
resumes to his previous activity.

Also fixes the issue reported in related bug on tablet devices:
the reconnection patch opens a second settings activity and if the user stays
and navigates into this second activity it could be prompted to redo the adapter
reconnection unexpectedly.

Fix Null Pointer Exception in Wifi Display Settings Reconnect dialog
This patch checks that activeDisplay is defined before using it.

Category: device enablement
Domain: CWS.WIFI-Common
Origin: internal
Upstream-Candidate: no, proprietary
Change-Id: I7d3327e6a2b985e74dec4285a98ccd44c9295ba1
Orig-Change-Id: I1fcd8058a23e2320e56becfa83f1f0e63b3adfe0
Orig-Change-Id: I9923748a3295a470f20bdf206794794747b5b389
Signed-off-by: Matthieu Mauger <matthieux.mauger@intel.com>