[CWS_WIDI_UX] Adapter reconnection upon AP just connected on another frequency
Since some WiFi chip have some limitation to offer high video quality while used
in muli-roles and multi-channels, the best solution is to use the same channel as
the access point when running in multi-roles.

This solution is already in place when DUT is first connected to an access point
and then connected to a video adapter. However, in some case we can imagine that
user first connects to video adapter and then connects to an access point. In
such situation, there is a high probability that adapter and access point
doesn't run on the same channel.

This patch detects when DUT is reaching an access point with a different
frequency than video adapter, a dialog box informs the user that the video link
will be closed and restored for better quality. The user has no other choice
than accept and the patch manages the disconnection, retrieves the adapter we
left and finally performs the reconnection.

Closes the Wireless Display Setting window as soon as the reconnection process
is completed. By the way, the user resumes to his previous activity.

Change-Id: I7225b5dba19ca403df2bf0a2de53339e5addc15f
Category: aosp improvement
Domain: CWS.WIFI-Common
Origin: internal
Upstream-Candidate: yes
Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-11392
Signed-off-by: Matthieu Mauger <matthieux.mauger@intel.com>
Signed-off-by: Paul Naoumenko <paulx.naoumenko@intel.com>
Signed-off-by: Cedric IZOARD <cedricx.izoard@intel.com>
Signed-off-by: Jean Trivelly <jean.trivelly@intel.com>
Signed-off-by: Aymen Ben Alaya <aymenx.ben.alaya@intel.com>
Reviewed-on: https://android.intel.com:443/307451