[PORT FROM MAIN] WiFi Display: Adapter reconnection upon AP just connected on another frequency
BZ: 113453 122779

Since Broadcom chip has some limitation to offer high video quality while used
in muli-role and multi-channel, the best solution is to use the same channel as
the access point when running in multi-role.

This solution is already in place when DUT is first connected to an access point
and then connected to a video adapter. However, in some case we can image that
user first connects to video adapter and then connects to an access point. In
such situation, there is a  high probability that adapter and access point
doesn't run on the same channel.

This patch detects when DUT is reaching an access point with a different
frequency than video adapter, a dialog box informs the user that the video link
will be closed and restored for better quality. The user has no other choice
than accept and the patch manages the disconnection, retrieves the adapter we
left and finally performs the reconnection.

Closes the Wireless Display Setting
window as soon as the reconnection process is completed. By the way, the user
resumes to his previous activity.

Also fixes the issue reported in related bug on tablet devices:
the reconnection patch opens a second settings activity and if the user stays
and navigates into this second activity it could be prompted to redo the adapter
reconnection unexpectedly.

Category: AOSP improvement
Domain: WIFI-Core
Severity: high
Origin: internal
Upstream-Candidate: no

Change-Id: I1fcd8058a23e2320e56becfa83f1f0e63b3adfe0
Orig-Change-Id: I72fb7ca84382cb246779a7eb4ef0b9ebe01b95d0
Orig-Change-Id: Iea9b64f6f48d41f8e42a29ce5dc7e26564fd395e
Signed-off-by: Matthieu Mauger <matthieux.mauger@intel.com>
Reviewed-on: http://android.intel.com:8080/122821
Reviewed-by: buildbot <buildbot@intel.com>
Reviewed-by: Bachot, Jean-michel <jean-michel.bachot@intel.com>
Reviewed-by: Naoumenko, PaulX <paulx.naoumenko@intel.com>
Tested-by: Champciaux, Nicolas <nicolas.champciaux@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>