WiFi Display: resume to previous activity after adapter reconnection
BZ: 122779

Recently, a patch has been merged to invite the user to reconnect to the video
adapter upon AP connection on a different channel:
    ID: I72fb7ca84382cb246779a7eb4ef0b9ebe01b95d0
    Author: Matthieu Mauger <matthieux.mauger@intel.com>
    Date: Jun 27, 2013

This new patch enhances the previous one by closing the Wireless Display Setting
window as soon as the reconnection process is completed. By the way, the user
resumes to his previous activity.

This enhancement also fixes the issue reported in related bug on tablet devices:
the reconnection patch opens a second settings activity and if the user stays
and navigates into this second activity it could be prompted to redo the adapter
reconnection unexpectedly.

Change-Id: Iea9b64f6f48d41f8e42a29ce5dc7e26564fd395e
Signed-off-by: Matthieu Mauger <matthieux.mauger@intel.com>
Reviewed-on: http://android.intel.com:8080/118401
Reviewed-by: Garcia, JeremieX <jeremiex.garcia@intel.com>
Reviewed-by: Trivelly, Jean <jean.trivelly@intel.com>
Reviewed-by: Chen, HuaqiangX <huaqiangx.chen@intel.com>
Reviewed-by: Naoumenko, PaulX <paulx.naoumenko@intel.com>
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Reviewed-by: Becle, MichelX <michelx.becle@intel.com>
Tested-by: Lestienne, ChristopheX <christophex.lestienne@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>