[PORT FROM MAIN] Wifi: Change label of Tethering settings shortcut
BZ: 123770

In the settings UI, the label for tethering is processed based on the regexps defined in the overlays.
Shortcut for Tethering is fixes to tethering & portable hotspot but should be only tethering, as portable hotspot
has it own separate widget.

Category : AOSP Improvement
Domain: WIFI-Settings
Origin: internal
Upstream-Candidate: no

Change-Id: I215ad3c677e3294d98f6e1d2345cf8f966af537a
Orig-Change-Id: I2d56b8c2e5d4514c6cc13fb972222264389ad6ac
Signed-off-by: Paul Naoumenko <paulx.naoumenko@intel.com>
Reviewed-on: http://android.intel.com:8080/123810
Reviewed-by: Poulain, LoicX <loicx.poulain@intel.com>
Tested-by: Champciaux, Nicolas <nicolas.champciaux@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>