Wifi Display: use reconnectWifiDisplay
BZ: 124677

Currently the reconnection (ie. disconnection + wait scan + connection)
to a WifiDisplay is done in the WifiDisplay Setting activity. Thus if user
is leaving the activity during the wait_scan step, the reconnection never happens.

This patch use the reconnectWifiDisplay (cf depending patch) interface instead,
so that as soon as user click ok, the reconnection will happen even if he quit the activity.

Change-Id: Ie4b0c711741d39adc81ea2750e971ffff2929078
Depends-On: I069a244c2e6de6753f97c646d49b5eb4cfc43b75
Signed-off-by: Cedric IZOARD <cedricx.izoard@intel.com>
Reviewed-on: http://android.intel.com:8080/121846
Reviewed-by: buildbot <buildbot@intel.com>
Reviewed-by: Mauger, MatthieuX <matthieux.mauger@intel.com>
Reviewed-by: Bachot, Jean-michel <jean-michel.bachot@intel.com>
Reviewed-by: Chen, HuaqiangX <huaqiangx.chen@intel.com>
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Tested-by: Lestienne, ChristopheX <christophex.lestienne@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>