[PORT FROM MAIN JB-MR2]wifi: avoid reconnection dialog to appear twice
BZ: 128660

When the DUT is connecting to an AP while being connected to a Wifi display
adapter, the user is prompted to perform a reconnection with the adapter.
However, if for any reason this reconnection fails, the dialog box can reappears
unexpectedly while user navigates in Settings application on a tablet form
factor.

This patch clears the Extra in launcher Intent as soon as the user accepts the
reconnection. This way, the dialog not appears again whatever the result of the
reconnection.

Category: device enablement
Domain: CWS.WIFI-Common
Origin: internal
Upstream-Candidate: no, proprietary
Change-Id: I38300e6b247e331c87b3612ceaa1d94ae04a9e52
Orig-Change-Id: I10e49ce4fc8a647524d965046010cf8f8d12aabc
Signed-off-by: Matthieu Mauger <matthieux.mauger@intel.com>
Reviewed-on: http://android.intel.com:8080/126442
Reviewed-by: Champciaux, Nicolas <nicolas.champciaux@intel.com>
Reviewed-by: Chen, HuaqiangX <huaqiangx.chen@intel.com>
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: Champciaux, Nicolas <nicolas.champciaux@intel.com>
Tested-by: cactus <cactus@intel.com>