WiFi Display: avoid reconnection dialog to appear twice
BZ: 128660

When the DUT is connecting to an AP while being connected to a Wifi display
adapter, the user is prompted to perform a reconnection with the adapter.
However, if for any reason this reconnection fails, the dialog box can reappears
unexpectedly while user navigates in Settings application on a tablet form
factor.

This patch clears the Extra in launcher Intent as soon as the user accepts the
reconnection. This way, the dialog not appears again whatever the result of the
reconnection.

Category: aosp improvement
Domain: WIDI-Settings
Origin: internal
Upstream-Candidate: no, proprietary

Change-Id: I02c73459e72fd600d5a97961d325d38b0b2f5f7b
Signed-off-by: Matthieu Mauger <matthieux.mauger@intel.com>
Reviewed-on: http://android.intel.com:8080/126440
Reviewed-by: Champciaux, Nicolas <nicolas.champciaux@intel.com>
Tested-by: Champciaux, Nicolas <nicolas.champciaux@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>