WiFi Display: keep adapters list when screen orientation changes
BZ: 125277

Until now, the P2P scan was stopped when leaving the Wireless Display Settings
page.  However when screen orientation change the Android system first closes
the page before recreate it with the new orientation. This behavior makes the
adapter list to be empty immediately after rotation and the user must wait for
adapters being discovered again by a new scan.

This patch detects if the Wireless Display Settings page is left definitively or
just for screen rotation. In the first case the scan is stopped as before but in
the second case the scan and scan results are kept in order to recover
everything as soon as the page has been recreated with the new orientation.

Change-Id: I9256c3d029fafcdca8b46fc279243fe273e31b30
Signed-off-by: Matthieu Mauger <matthieux.mauger@intel.com>
Reviewed-on: http://android.intel.com:8080/125275
Tested-by: Sun, Hang L <hang.l.sun@intel.com>
Reviewed-by: Bachot, Jean-michel <jean-michel.bachot@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>