[PORT FROM R42B-STABLE] WiFi P2P: handle FileTransferServerService restart
BZ: 134124

When system memory is running low, Android can kill some services and restarts
them as soon as memory resumes to a normal level. This is what happened in bug
referenced here.

In such case, Android restarts the service by calling onStartCommand with null
argument, we must handle this situation and this is the aim of this patch.

Category: AOSP Improvement
Domain: WIFI-Settings
Origin: internal
Upstream-Candidate: no, proprietary

Change-Id: I270b240d36fd1c6594f00681a3afbd7bbdc0996f
Orig-Change-Id: I968570560906b8027e5b5045bd496481c4df5911
Signed-off-by: Matthieu Mauger <matthieux.mauger@intel.com>
Reviewed-on: http://android.intel.com:8080/129086
Reviewed-by: Samoun, JacquesX <jacquesx.samoun@intel.com>
Reviewed-by: Chen, HuaqiangX <huaqiangx.chen@intel.com>
Reviewed-by: Yang, Bin Y <bin.y.yang@intel.com>
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Reviewed-by: Naoumenko, PaulX <paulx.naoumenko@intel.com>
Reviewed-by: Bachot, Jean-michel <jean-michel.bachot@intel.com>
Tested-by: Lestienne, ChristopheX <christophex.lestienne@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>