WiFi: Fix stuck Hotspot  when selecting 5GHz band and Auto Channel
BZ: 136193

When Setting Hotspot to 5GHz band and selection Auto Channel, the
Hotspot automatically turns off and could not be enabled any more.

The problem have been seen only with a chineese SIM card. It turns
out that when selecting auto Channel, a harcoded channel (36) was
set regardless of the Country Code. This leads in a stuck when the
channel 36 is not supported in Hotspot feature of the country.
The issue was fixed by getting the channel from  the first element
of the channel list when auto feature is set.

Category: aosp improvement
Domain: WIFI Settings
Origin: internal
Upstream-Candidate: yes

Change-Id: I788c65d0a388b425036fdd16d2eba3d16c9a61bb
Signed-off-by: Aymen KLOUZ <aymenx.klouz@intel.com>
Reviewed-on: http://android.intel.com:8080/130833
Reviewed-by: Izoard, CedricX <cedricx.izoard@intel.com>
Reviewed-by: Trivelly, Jean <jean.trivelly@intel.com>
Reviewed-by: Chen, HuaqiangX <huaqiangx.chen@intel.com>
Reviewed-by: Yang, Bin Y <bin.y.yang@intel.com>
Tested-by: Lestienne, ChristopheX <christophex.lestienne@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>