WiFi Settings: Align EAP methods for all localizations
BZ: 142765

EAP methods are stored in a strings array. This array was not aligned for all
localizations so depending of the set language, the number of EAP methods was
not the same. Only the default one (US) was complete.

These patch aligns the methods array all over the localizations.

Category: aosp improvement
Domain: CWS.WIFI-Common
Origin: internal
Upstream-Candidate: no, proprietary

Change-Id: I6a5959272b5f412818832fdcf40538c4e5b30c6d
Signed-off-by: Matthieu Mauger <matthieux.mauger@intel.com>
Reviewed-on: http://android.intel.com:8080/138174
Reviewed-by: Trivelly, Jean <jean.trivelly@intel.com>
Reviewed-by: Garcia, Jeremie <jeremie.garcia@intel.com>
Reviewed-by: Klouz, AymenX <aymenx.klouz@intel.com>
Reviewed-by: Bachot, Jean-michel <jean-michel.bachot@intel.com>
Reviewed-by: Yang, Bin Y <bin.y.yang@intel.com>
Reviewed-by: Chen, HuaqiangX <huaqiangx.chen@intel.com>
Reviewed-by: Yin, Fengwei <fengwei.yin@intel.com>
Tested-by: Lestienne, ChristopheX <christophex.lestienne@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>