ntegrate Arkham to mainline
BZ: 156819

Arkham is a deeply integrated security container capability for
Android for the purpose of isolation and management in mixed used
cases. Primarily Arkham is intended to enable separation of work and
personal content and capability.

Detailed information about the Arkham requirements and architecture can
be found here: https://sharepoint.amr.ith.intel.com/sites/SDI/Arkham/

This patch modifies the Setting application to implement the following
Arkham features:

* Warns users that activating the USB debugging will lock the
container.

* Modifies the Keyguard screen shown to the user when unlocking a
container to display the container name and container icon

* Allow installation into Credential Storage for container
users. This is required so that private corporate certificates can be
installed in the container and thus not available to primary user.

* Remove the brightness and wallpaper settings from the container
Settings application since these settings are common with the primary
user and it is confusing to show to the user the same settings in two
places.

* Remove the sync account type settings. As contacts can be merged
with the primary user based on policy set by the container
administrator, this setting should not be visible to the user.

* Running container applications should be displayed in the Settings
Running Application panel for the user who created the container.

* Hide container users in the Settings Users panel.

* Fix JIRA: ARKHAM-1268
The reported Klockwork issues:
https://android-kw2.jf.intel.com:8090/review/insight-review.html#goto:project=mainline_redhookbay,pid=107398
https://android-kw2.jf.intel.com:8090/review/insight-review.html#goto:project=mainline_redhookbay,pid=107399
https://android-kw2.jf.intel.com:8090/review/insight-review.html#goto:project=mainline_redhookbay,pid=107410

* Fix crash when trying to access running apps from Settings
Search in disabled packages when getting the service info in
updateService()

Change-Id: I8789a0186e81a3931548816742e83d3434e514ab
Category: feature differentiation
Domain: AOSP.Security-Arkham
Origin: internal
Upstream-Candidate: no, proprietary
Signed-off-by: Octavian Purdila <octavian.purdila@intel.com>
Reviewed-on: http://android.intel.com:8080/126925
Reviewed-by: Yang, Bin Y <bin.y.yang@intel.com>
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Reviewed-by: cactus <cactus@intel.com>