wifi: Update AP status if association is rejected
BZ: 154541

If a remembered AP, is configured to deny assocication to DUT
(with MAC filtering for example), after several attempts to connect
the AP ends up at the top of AP list without status.

The case when AP connection failed because of association is
rejected was not handle when updating AP status.
The patch handles this case just like an Authentication failure

Change-Id: Iaccd7ed6c6abdd8fda09c7369592297e04e62cf8
Category: device enablement
Domain: CWS.WIFI-Common
Origin: internal
Upstream-Candidate: yes
Signed-off-by: Cedric IZOARD <cedricx.izoard@intel.com>