[MERR_PR2-Kitkat] There is not advanced options to Set the Soft AP Frequency band and channel in Hotspot Setting.
BZ: 154138

SoftAP under KitKat does not have the option to change the band frequency
nor the channel to use.
This patch add this possibility into the Settings. The user can change the band
(2.4 or 5 Ghz), change the hw mode 'bg','a' or 'ac'), set the 80211 mode and
select radio channel according to local regulatory.
When the Soft Ap is started and the advanced menu is activated, the settings
(band frequency, channels, hw mode) will be provided by the chipset and displayed
accordingly for the user to chose.
When the Soft Ap is not started and the advanced menu is activated, the settings
will be by defaulted to band 2.4Ghz only, and channel choice from 1 to 11.

Change-Id: Ica16214931c3a102774c5acf2fd8bc08813e3405
Category: device enablement
Domain: AOSP.Connectivity-Ethernet-settings
Origin: internal
Upstream-Candidate: no, proprietary
Signed-off-by: Marc Blassin <marcx.blassin@intel.com>
Signed-off-by: Paul Naoumenko <paulx.naoumenko@intel.com>
Signed-off-by: Jean-Michel Bachot <jean-michel.bachot@intel.com>