[PORT FROM R43] Settings: SIM Card Option doesn't reappear when SIM is inserted
BZ: 106334

When a SIM PIN enabled SIM is removed and inserted while in setup SIM Card lock
screen, after entering the PIN the option doesn't reappear.

The reason for this is, when the SIM is removed while in Setup SIM Card Lock screen
the screen is dimmed, when the user presses back the resume of Security Settings is
called and since there is NO SIM the SIM LOCK option is removed. But when the SIM
is re-inserted, the updatesimpreferences, finds that there is no SIM PIN option.

This patch fixes this issue by calling the createPreferenceHierarchy, if there is
no SIM LOCK option and the device has an Icc card.

Change-Id: I8bec14598dd0d46e9c7150f9a9a14c9365b9b093
Orig-Change-Id: I226ad64a86918968b09a8ea0412b4b28f5f4a166
Category: aosp improvement
Domain: Telephony-Settings
Origin: internal
Upstream-Candidate: yes
Signed-off-by: Arun Ravindran <arun.ravindran@intel.com>