LPAL: Reset voice unlock flag when start face unlock
BZ: 178062

When starts face unlock, it passes a PendingIntent as Intent extra to
the face unlock app. But if voice unlock is first set up and canceled,
a flag in the PendingIntent "LOCKSCREEN_BIOMETRIC_VOICE_WEAK_FALLBACK"
is set to "true". Then later, when starts face unlock, it gets this
"stored" PendingIntent with this incorrect flag set, which result in the
face unlock is finally set as voice unlock in setting menu.

Change-Id: I42e82ba09b02cffc93b9faf51746189c2585efb4
Category: feature differentiation
Domain: LPAL-Common
Origin: internal
Upstream-Candidate: no, proprietary
Signed-off-by: Jian, Qinghui <qinghui.jian@intel.com>