WIFI: EAP-FAST: set a unique PAC name per EAP-FAST network
BZ: 196917

Currently a single pac name is use for all EAP-FAST secured networks.
This leads to connection issue because PAC contains and encrypted key
delivered by the server. So the PAC is created for the first connection
to first server and leads to connection failures for other networks.
In addition as PAC is based on phase2 information, modify the phase2
of a network is not possible.

So each network configuration should have it's own PAC.
Fix it by adding to the PAC name a suffix based on phase2 method, ssid and Identity.
This fix permits to  :
- Have multiple networks registered and switch between them.
- Modify phase2 is now possible.
- forget then re-create a neworks will re-use the previous PAC.

Change-Id: Ic41ef20241cd7945f7161d6cd06d23766584e1bf
Category: aosp improvement
Domain: CWS.WIFI-Common
Origin: internal
Upstream-Candidate: yes
Signed-off-by: Honore Tricot <honorex.tricot@intel.com>