[PATCH] Settings: The cellular data button state is wrong
While turning on the cellular data button with the changing of
default data sim, the button state may not switch to "on".
The root cause is that the sequence in time is wrong, so when it
used the method of updatePolicy, it can't get the right state.
Now we use ContentObserver to avoid it.

Change-Id: Ib678eaeb6bdf3460598b3dfa940366134e100c2f
Category: AOSP improvement
Domain: Settings-Applications
Origin: Internal
Upstream-Candidate: no
Signed-off-by: saishuax <saishuaix.qu@intel.com>