[BT] Tombstone when enabbling/disabling BT
The tombstone happens because some binder between BT
classes exists no more. This defect occurs when the user
send enabling and disabling command so quickly, that
the created object can't be deleted, and those deleted
are not created.

The patch prevent this defect limiting the inputs the
user can send. In effect, in R5, it is possible to
toggle the BT with the button or with the bar. Doing it
with the user bypass some protection of the button and
send a lot of ON/OFF request.

Change-Id: I9223335beb6c16526eef8189803c50868e911ac6
Category: aosp improvement
Domain: CWS.BT-Common
Origin: internal
Upstream-Candidate: yes
Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-27601
Reviewed-on: https://android.intel.com:443/355406