[CWS_EAP_SIM_AKA] Use ConnectivityManager to check telephony feature
PackageManager.FEATURE_TELEPHONY description is not clear, in
http://developer.android.com/reference/android/content/pm/PackageManager.html
this feature is present as soon as the device has a telephony radio with data
but in another document
http://static.googleusercontent.com/media/source.android.com/en/us/compatibility/5.0/android-5.0-cdd.pdf
The same feature is described as a SMS/Voice capable device.
This leads to confusion on devices that are data only capable like
cherrytrail boards.

To fix that, this patch prefers to request NetworkInfo from
ConnectivityManager to check if device has a cellular connectivity
or not.

Change-Id: I58c14aec4969f07dda1b36de848eea1b53cb5563
Category: feature differentiation
Domain: CWS.WIFI-EAP_SIM_AKA
Origin: internal
Upstream-Candidate: no, proprietary
Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-28901
Signed-off-by: Matthieu MAUGER <matthieux.mauger@intel.com>
Reviewed-on: https://android.intel.com:443/359695