[PATCH] Settings: Preferred sending SMS messages SIM
When sending a sms, promise a identifier to judge if it is in
ask every time mode. It can be set in Settings, and if it is in
ask every time mode, Messaging will pop a dialog to make the user
choose a sim to send sms. Telephony will send the sms using the subid
on the basis of the user choose.

Change-Id: I4c5dc3b4b54231e590c891c9886c30b8fdf3660f
Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-31113
Category: AOSP improvement
Domain:  Settings
Origin: Internal
Upstream-Candidate: no
Signed-off-by: jialinx.dong <jialinx.dong@intel.com>